feat: 🎸 Add org wide pull request template (#1)
* feat: 🎸 Add org wide pull request template

Adding it here will enable it org wide, meaning if we want to change it
in the future it will be less effort than having to do it in a bunch of
repos. After doing this here and merging it, I will go back and remove
all repo specific ones so they fallback to using this universal one so
it's all standardised. I tried to find the simple one @des-des did but
couldn't find it, so whipped this up real quick. I'm happy to
incorporate Eoin's one, or replace this entirely. It's 100% comment so
it's less intrusive and doesn't interrupt commit message heavy workflows
(like mine), while still prompting for important information. Tried to
keep it as open as possible while still making sure to prompt for all
the key bits.

* docs: Update pr template notes on terraform